move content related to 'MapComponent' from 'HomeComponent' to dedicated 'MapPage'
